checkout fixes from master (0.9.0 essentially) to back-port into v0.7.2 that works on py3.6 in VM

+++ mesmerize/viewer/modules/batch_run_modules/CNMFE.py
@@ -207,11 +207,12 @@ def run(batch_dir: str, UUID: str):
 
     dview.terminate()
 
-    for mf in glob(batch_dir + '/memmap-*'):
-        try:
-            os.remove(mf)
-        except: # Windows doesn't like removing the memmaps
-            pass
+    if not input_params.get('keep_memmap', False):
+        for mf in glob(batch_dir + '/memmap-*'):
+            try:
+                os.remove(mf)
+            except: # Windows doesn't like removing the memmaps
+                pass
 
     end_time = time()
     processing_time = (end_time - start_time) / 60
