+++ mesmerize/viewer/modules/batch_run_modules/caiman_motion_correction.py
@@ -80,7 +80,7 @@ def run_single(batch_dir, UUID, output):
 
     fname = [file_path + '_input.tiff']
     # TODO: Should just unpack the input params as kwargs
-    input_params = pickle.load(open(file_path + '.params', 'rb'))
+    input_params: dict = pickle.load(open(file_path + '.params', 'rb'))
     mc_kwargs = input_params['mc_kwargs']
 
     splits_rig = n_processes
@@ -125,16 +125,17 @@ def run_single(batch_dir, UUID, output):
         m_els = m_els.astype(np.uint16)
 
     img_out_path = os.path.join(batch_dir, f'{UUID}_mc.tiff')
-    tifffile.imsave(img_out_path, m_els, bigtiff=True, imagej=True, compress=1)
+    tifffile.imsave(img_out_path, m_els, bigtiff=True, imagej=False, compress=1)
     output['output_files'] = [UUID + '_mc.tiff']
 
     output.update({'status': 1, 'bord_px': int(bord_px_els)})
 
-    for mf in glob(os.path.join(batch_dir, UUID + '*.mmap')):
-        try:
-            os.remove(mf)
-        except:
-            pass
+    if not input_params.get('keep_memmap', False):
+        for mf in glob(os.path.join(batch_dir, UUID + '*.mmap')):
+            try:
+                os.remove(mf)
+            except:
+                pass
 
     dview.terminate()
 
@@ -214,7 +215,7 @@ def run_multi(batch_dir, UUID, output):
         mc_out[:, z, :, :] = m_els
 
     img_out_path = os.path.join(batch_dir, f'{UUID}_mc.tiff')
-    tifffile.imsave(img_out_path, mc_out, bigtiff=True, imagej=True, compress=1)
+    tifffile.imsave(img_out_path, mc_out, bigtiff=True, imagej=False, compress=1)
     output['output_files'] = [f'{UUID}_mc.tiff']
 
     output.update(
